Rename basic_types.py to types.py (types_test.py still works via pytest invocation)
